Working version

+++ GPIO/gpio.h
@@ -1,5 +1,5 @@
 void pinMode (int pin, char* MODE);
 void digitalWrite(int pin, int value);
 int digitalRead(int pin);
-void blink(int pin, int freq, int duration);
+void blink(int pin, float freq, int duration);
